+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-05-24 19:12:50
-// Commit: refactor(test): optimize data flow between domain and data layers
+// Updated at: 2026-05-25 14:26:53
+// Commit: feat(test): integrate custom telemetry logger in test
 
 class testComponentState {
-  final String commitRef = '1414';
-  final String lastMessage = 'refactor(test): optimize data flow between domain and data layers';
-  final DateTime timestamp = DateTime.parse('2026-05-24T19:12:50+05:00');
+  final String commitRef = '1429';
+  final String lastMessage = 'feat(test): integrate custom telemetry logger in test';
+  final DateTime timestamp = DateTime.parse('2026-05-25T14:26:53+05:00');
 }